Add shared Category Options
